Introduce flags and case sensitivity.
This change introduces flags `(?...)` in glob expressions. Flags are
used exclusively to toggle case sensitivity with the `i` flag. Flags
mutate state in the parser and are not emitted as tokens.

`Glob::partitioned` and other APIs now consider the case sensitivity of
literals and the target platform's file system as well as the contents
of literals when constructing a literal path prefix and partitioning a
glob expression. If a literal sequence contains any characters with
casing and any configured case sensitivity that disagrees with the
target platform's file system, then the literal is considered variant.

Globs now default to the case sensitivity of the target platform's file
system.