Refactor parsing and parser errors.
This change integrates with `nom_supreme` and fixes several bugs in
`parse`. The `glob` and `wildcard` parsers are now configured with
contextual terminators, eliminating the need to leak terminators between
parsers and the need for `no_adjacent_tree`. Instead, the acceptable
terminators for tree tokens (and others) are well defined within the
parsing function. Parser inputs are always located and this information
is used to detect the beginning of glob sub-expressions to further
delimit tree tokens.

Errors are now based on `ErrorTree`, which provides more instrumentation
than `nom`'s default error types. This is used to report the start of a
parser error as well as errors that occurred deepest in the parse.