Provide an `any` combinator for matching.
This change introduces the `Any` type and `any` combinator function.
`Any` is a combinator that trivially packs token sequences into a parent
`Alternative`. The alternative token forms a set of globs that can be
matched together ergonomically and efficiently. `Any` does not apply the
same rules as `Glob` and allows for rooted and unrooted patterns.

Matching has been generalized by the `Pattern` trait, which must now be
imported to use the `is_match` and `matched` functions. Both `Glob` and
`Any` implement `Pattern`, but `Any` does not provide a `walk` function
by design. The `any` function can compose any `Pattern` type.