Implement boundedness and variance queries.
This change introduces boundedness and variance of tokens. Boundedness
describes the depth and breadth of expressions with respect to a
directory tree. Variance is more sophisticated than a binary query, and
describes any invariant string representation and the boundedness of
variance.

These concepts are used to more robustly detect terminal negations,
including expressions like `a/<<?>/>*`. Previously, only expressions
terminated with a tree token were considered terminal.

A simplified `Variance` type is exposed in the public API and
`is_invariant` has been replaced by `variance`. For invariant globs,
this exposes an equivalent `PathBuf`.