Do not read directory trees that match terminal negations.
This change refactors `Walk` and supporting code out of the crate root
and into a `walk` module. Negations are now partitioned into terminal
and nonterminal patterns. When a terminal pattern is matched, the
traversal does not descend into sub-trees. This avoids reading
potentially large trees when it is not necessary. Nonterminal patterns
(those that do not end with a tree wildcard) are used as a simple
per-entry filter.