Consider paths when components and patterns are exhausted.
This change considers paths when walking a directory tree even if there
are no initial path components nor pattern components after stripping
the prefix and depth. This has two primary effects: when a glob consists
solely of a tree wildcard, the root of the tree is included in the
output and literal paths are matched correctly when partitioning an
invariant glob expression.

See https://github.com/olson-sean-k/wax/issues/11.