Accept types with compatible errors in `Walk::not`.
The constraints on the pattern type `P` in `Walk::not` were too strict.
This change accepts any type that when converted into a `Glob` emits an
error that can be converted into `BuildError`. This is important for
supporting glob parameters (where `P` is `Glob<'_>`) where the error
type is `Infallible` and not `BuildError`.