Unconditionally provide located errors.
This change refactors the `diagnostics` module and its associated
feature such that error locations are unconditionally included in
`BuildError`s. The `diagnostics` feature has been renamed to `miette`
and is only concerned with integration with the `miette` crate and
related libraries; basic diagnostics are always present and exposed via
the `BuildError::locations` function.

Error locations are exposed via `LocatedError` trait objects, which
provide an associated `Span` and a `Display` implementation. This trait
is used to expose error details for parse and rule failures.

As part of this change, the MSRV has been bumped to `1.60.0` to take
advantage of the `dep:` prefix in the Cargo manifest.