Rename `Pattern` to `Program` and `Combine` to `Pattern`.
This change renames the `Pattern` and `Combine` traits. The `Combine`
trait is used in `any` APIs, but it does not principally concern
combinators. Note too that `Pattern: Combine`. `Combine` broadly
describes pattern representations and, more specifically, types that can
be converted into a checked token tree. `Pattern` is more specific, and
describes _compiled_ pattern representations that can be matched against
paths.

As such, `Combine` is now `Pattern` and `Pattern` is now `Program`. This
actually makes `any` APIs less confusing, as they accept `Pattern`s,
whatever there form may be.

This change also alters the nomenclature in documentation and the
conceptualization of `Combine`.