Implement decomposition into alternatives and flatten APIs.
This change flattens APIs that accept sequences of patterns in favor of
accepting a single pattern and instead relying on combinators (i.e.,
`wax::any`). The only such API is `FileIterator::not`, which accepted a
sequence so that patterns can be partitioned by exhaustiveness and
exhaustive matches can be detected. This avoids unnecessary reads.

However, this presents a less consistent API and is limited to the
alternatives provided explicitly by callers. This change implements
decomposition of a token into alternatives, which is more reliable and
flexible. In `FileIterator::not`, the pattern is decomposed and
subsequently partitioned.