Expose diagnostic and inspection types in a `query` module.
This change introduces a `query` module for types and APIs that diagnose
and inspect patterns and programs. Both depth and text variance are now
exposed by the `Program` trait using an API that is similar to internal
APIs but is simplified.

Additionally, `GlobVariance` has been renamed to `TokenVariance` and
type aliases for natural bound types are used more consistently.